Conflicts:
	cli/src/main/java/hudson/cli/CliPort.java
	cli/src/main/java/hudson/cli/Connection.java
	cli/src/test/java/hudson/cli/ConnectionTest.java
	core/src/main/java/hudson/DNSMultiCast.java
	core/src/main/java/hudson/DependencyRunner.java
	core/src/main/java/hudson/cli/BuildCommand.java
	core/src/main/java/hudson/cli/CommandDuringBuild.java
	core/src/main/java/hudson/cli/GroovyCommand.java
	core/src/main/java/hudson/cli/declarative/CLIRegisterer.java
	core/src/main/java/hudson/console/AnnotatedLargeText.java
	core/src/main/java/hudson/console/ConsoleAnnotatorFactory.java
	core/src/main/java/hudson/diagnosis/HudsonHomeDiskUsageMonitor.java
	core/src/main/java/hudson/lifecycle/UnixLifecycle.java
	core/src/main/java/hudson/logging/LogRecorder.java
	core/src/main/java/hudson/logging/LogRecorderManager.java
	core/src/main/java/hudson/model/AdministrativeMonitor.java
	core/src/main/java/hudson/model/AperiodicWork.java
	core/src/main/java/hudson/model/AsyncAperiodicWork.java
	core/src/main/java/hudson/model/FullDuplexHttpChannel.java
	core/src/main/java/hudson/model/HealthReport.java
	core/src/main/java/hudson/model/JobPropertyDescriptor.java
	core/src/main/java/hudson/model/ParametersDefinitionProperty.java
	core/src/main/java/hudson/model/Resource.java
	core/src/main/java/hudson/model/ViewGroupMixIn.java
	core/src/main/java/hudson/model/queue/AbstractQueueTask.java
	core/src/main/java/hudson/model/queue/MappingWorksheet.java
	core/src/main/java/hudson/node_monitors/AbstractAsyncNodeMonitorDescriptor.java
	core/src/main/java/hudson/node_monitors/ResponseTimeMonitor.java
	core/src/main/java/hudson/os/solaris/ZFSInstaller.java
	core/src/main/java/hudson/scm/PollingResult.java
	core/src/main/java/hudson/security/BasicAuthenticationFilter.java
	core/src/main/java/hudson/security/HudsonAuthenticationEntryPoint.java
	core/src/main/java/hudson/tasks/CommandInterpreter.java
	core/src/main/java/hudson/triggers/SafeTimerTask.java
	core/src/main/java/hudson/util/AlternativeUiTextProvider.java
	core/src/main/java/hudson/util/ArgumentListBuilder.java
	core/src/main/java/hudson/util/CyclicGraphDetector.java
	core/src/main/java/jenkins/ExtensionFilter.java
	core/src/main/java/jenkins/model/IdStrategy.java
	core/src/main/java/jenkins/model/lazy/LazyBuildMixIn.java
	core/src/main/java/jenkins/security/ConfidentialStore.java
	core/src/main/java/jenkins/security/MasterToSlaveCallable.java
	core/src/main/java/jenkins/security/QueueItemAuthenticatorConfiguration.java
	core/src/main/java/jenkins/security/SlaveToMasterCallable.java
	core/src/main/java/jenkins/security/s2m/MasterKillSwitchConfiguration.java
	core/src/main/java/jenkins/tasks/SimpleBuildStep.java
	core/src/main/java/jenkins/util/xstream/XStreamDOM.java
	core/src/main/resources/hudson/cli/CLIAction/command.jelly
	core/src/test/java/hudson/os/SUTester.java
	test/src/test/groovy/jenkins/model/BuildDiscarderTest.groovy
	test/src/test/java/hudson/console/ExceptionAnnotationTest.java
	test/src/test/java/hudson/console/UrlAnnotatorTest.java
	test/src/test/java/hudson/logging/LogRecorderManagerTest.java
	test/src/test/java/hudson/model/UpdateCenter2Test.java
	test/src/test/java/hudson/tasks/FingerprinterTest.java
	test/src/test/java/hudson/tools/ToolLocationNodePropertyTest.java
	test/src/test/resources/jenkins/model/BuildDiscarderTest/compatibility/config.xml
	test/src/test/resources/jenkins/model/BuildDiscarderTest/compatibility/jobs/foo/config.xml

+++ core/src/main/java/hudson/logging/LogRecorder.java
@@ -33,7 +33,6 @@ import hudson.model.*;
 import hudson.util.HttpResponses;
 import jenkins.model.Jenkins;
 import hudson.model.listeners.SaveableListener;
-import hudson.remoting.Callable;
 import hudson.remoting.Channel;
 import hudson.remoting.VirtualChannel;
 import hudson.slaves.ComputerListener;
@@ -157,16 +156,14 @@ public class LogRecorder extends AbstractModelObject implements Saveable {
 
         @Deprecated
         public boolean includes(LogRecord r) {
-            if(r.getLevel().intValue() < level) {
+            if(r.getLevel().intValue() < level)
                 return false;   // below the threshold
-            }
             if (name.length() == 0) {
                 return true; // like root logger, includes everything
             }
             String logName = r.getLoggerName();
-            if(logName==null || !logName.startsWith(name)) {
+            if(logName==null || !logName.startsWith(name))
                 return false;   // not within this logger
-            }
             String rest = logName.substring(name.length());
             return rest.startsWith(".") || rest.length()==0;
         }
@@ -177,9 +174,8 @@ public class LogRecorder extends AbstractModelObject implements Saveable {
                 return Boolean.valueOf(levelSufficient); // include if level matches
             }
             String logName = r.getLoggerName();
-            if(logName==null || !logName.startsWith(name)) {
+            if(logName==null || !logName.startsWith(name))
                 return null; // not in the domain of this logger
-            }
             String rest = logName.substring(name.length());
             if (rest.startsWith(".") || rest.length()==0) {
                 return Boolean.valueOf(levelSufficient); // include if level matches
@@ -199,9 +195,8 @@ public class LogRecorder extends AbstractModelObject implements Saveable {
          */
         public void enable() {
             Logger l = getLogger();
-            if(!l.isLoggable(getLevel())) {
+            if(!l.isLoggable(getLevel()))
                 l.setLevel(getLevel());
-            }
             new SetLevel(name, getLevel()).broadcast();
         }
 
@@ -296,15 +291,12 @@ public class LogRecorder extends AbstractModelObject implements Saveable {
         }
 
         List<Target> newTargets = req.bindJSONToList(Target.class, src.get("targets"));
-        for (Target t : newTargets) {
+        for (Target t : newTargets)
             t.enable();
-        }
         targets.replaceBy(newTargets);
 
         save();
-        if (oldFile!=null) {
-            oldFile.delete();
-        }
+        if (oldFile!=null) oldFile.delete();
         rsp.sendRedirect2(redirect);
     }
 
@@ -319,18 +311,15 @@ public class LogRecorder extends AbstractModelObject implements Saveable {
      */
     public synchronized void load() throws IOException {
         getConfigFile().unmarshal(this);
-        for (Target t : targets) {
+        for (Target t : targets)
             t.enable();
-        }
     }
 
     /**
      * Save the settings to a file.
      */
     public synchronized void save() throws IOException {
-        if(BulkChange.contains(this)) {
-            return;
-        }
+        if(BulkChange.contains(this))   return;
         getConfigFile().write(this);
         SaveableListener.fireOnChange(this, getConfigFile());
     }
@@ -344,14 +333,11 @@ public class LogRecorder extends AbstractModelObject implements Saveable {
         getParent().logRecorders.remove(name);
         // Disable logging for all our targets,
         // then reenable all other loggers in case any also log the same targets
-        for (Target t : targets) {
+        for (Target t : targets)
             t.disable();
-        }
-        for (LogRecorder log : getParent().logRecorders.values()) {
-            for (Target t : log.targets) {
+        for (LogRecorder log : getParent().logRecorders.values())
+            for (Target t : log.targets)
                 t.enable();
-            }
-        }
         rsp.sendRedirect2("..");
     }
